Refactored the app to accept decimal mg values for vitamins from the API

+++ LocalBarcodeDB.swift
@@ -14,7 +14,7 @@ struct LocalBarcodeDB {
         let code: String // UPC/EAN as digits, no spaces
 
         // Per-serving or per-package values (you decide how you populate JSON)
-        // Units: kcal for calories (Int), grams for macros (Double), mg for sodium and vitamins/minerals (Int, mg base).
+        // Units: kcal for calories (Int), grams for macros (Double), mg for sodium and vitamins/minerals.
         let calories: Int?
         let carbohydrates: Double?
         let protein: Double?
@@ -35,18 +35,18 @@ struct LocalBarcodeDB {
         let plantProtein: Double?
         let proteinSupplements: Double?
 
-        // Optional vitamins (mg base)
-        let vitaminA: Int?
-        let vitaminB: Int?
-        let vitaminC: Int?
-        let vitaminD: Int?
-        let vitaminE: Int?
-        let vitaminK: Int?
+        // Optional vitamins (mg base, now Double? to preserve decimals)
+        let vitaminA: Double?
+        let vitaminB: Double?
+        let vitaminC: Double?
+        let vitaminD: Double?
+        let vitaminE: Double?
+        let vitaminK: Double?
 
-        // Optional minerals (mg base)
+        // Optional minerals (mg base) — potassium now Double?
         let calcium: Int?
         let iron: Int?
-        let potassium: Int?
+        let potassium: Double?
         let zinc: Int?
         let magnesium: Int?
     }
